requirements.txt
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.0
decorator==4.0.10
docker-compose==1.8.1
docker-py==1.10.6
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
EasyProcess==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
hurry.filesize==0.9
ipaddress==1.0.17
jsonschema==2.5.1
junitxml==0.7
linecache2==1.0.0
nose==1.3.7
py==1.4.31
pytest==3.0.3
PyVirtualDisplay==0.2.1
PyYAML==3.12
requests==2.7.0
retry==0.9.2
robotframework==2.8.2
robotframework-databaselibrary==0.8.1
robotframework-selenium2library==1.8.0
sauceclient==0.2.1
selenium==3.0.1
singledispatch==3.4.0.3
six==1.10.0
sst==0.2.2
testrail==0.3.7
texttable==0.8.6
traceback2==1.4.0
unittest2==1.1.0
virtualenv==15.0.3
websocket-client==0.37.0
xvfbwrapper==0.2.8
